TVG-58 Move ShowEpisodes into components
